in process

+++ src/main/java/web/dao/UserDao.java
@@ -6,8 +6,8 @@ import java.util.List;
 
 public interface UserDao {
     List<User> getAllUsers();
-    User getUserById(long id);
+    User getUserById(int id);
     void saveUser(User user);
     void updateUser(User user);
-    void deleteUser(long id);
+    void deleteUser(int id);
 }
