+++ src/main/java/web/service/UserService.java
@@ -5,8 +5,8 @@ import java.util.List;
 
 public interface UserService {
     List<User> getAllUsers();
-    User getUserById(Long id);
+    User getUserById(int id);
     void saveUser(User user);
     void updateUser(User user);
-    void deleteUser(Long id);
+    void deleteUser(int id);
 }
